Began expanding content and rearranged some sections.
The ordering of the three sections of "basics, content, filtering"
seems off. If we think of this like we would when investigating
our logs, we should start with the basics, filter it down, then
look at the content. However, this doesn't blend well into the
GitHub section. I was planning to begin the GitHub section with
a similar "filtering" notion.

Signed-off-by: Brent Beer <brentbeer@github.com>